Revert "Add overlay component for job role selection"
